requirements.txt
RealtimeSTT
groq
python-dotenv
rich
requests
ddgs
python-docx
pypdf
yt-dlp
psutil
wikipedia
speedtest-cli
prompt_toolkit
pytz
ffmpeg
piper-tts